Change sign of mux input based on the "VCO sign"
Ties the sign of the VCO selected in the GUI (the sign of reference_frequency1 on the FPGA) to the sign of inst_frequency0 and pll0_output sent to the multiplexer. Additionally, resize_with_saturation.vhd was altered to avoid the "most negative number" of the signed 2's complement representation. Negating the most negative number returns the most negative number.