New Loop Filter Division Factors
Scaled the division factors going into the loop filters. The intent was to use the division factor that gave the smallest magnitude output when the error was maximized and the gain was minimized (to be able to reach the "weakest" feedback possible by minimizing the gain). The double integrator was scaled so that with maximum gain and maximum error the output will rail in 1us. With maximum error and minimum gain the double integrator should not rail for a few thousand seconds.